// app/(saas)/usecases/[id]/types/usecase.ts
export interface ScoreCategory {
  category: string
  color: string
  description: string
  icon: string
}